+++ src/Constants/Numbers.h
@@ -36,6 +36,7 @@ namespace constants
 
 	const int DEFAULT_STARTING_MORALE = 600;
 
-	const float ROBOT_COM_RANGE = 15.0;
-	const float COMM_TOWER_BASE_RANGE = 10.0;
+	const float ROBOT_COM_RANGE = 15.0f;
+	const float COMM_TOWER_BASE_RANGE = 10.0f;
+	const float LANDER_COM_RANGE = 5.0f;
 }