+++ src/Constants/Strings.h
@@ -239,6 +239,7 @@ namespace constants
 
 	const std::string ALERT_LANDER_TILE_OBSTRUCTED = "Cannot place Lander because there is an object on the selected tile.";
 	const std::string ALERT_LANDER_TERRAIN = "Landers cannot be placed on Impassable Terrain.";
+	const std::string ALERT_LANDER_COMM_RANGE = "Landers must be placed within 5 tiles of the Command Center.";
 
 	const std::string ALERT_STRUCTURE_OUT_OF_RANGE = "Cannot build structures more than " + std::to_string(constants::ROBOT_COM_RANGE) + " tiles away from Command Center.";
 	const std::string ALERT_STRUCTURE_TILE_OBSTRUCTED = "The selected tile already has a structure on it. You must bulldoze the existing structure in order to build here.";